moving MAR code to shared folder
